area-types handlers in separate file, tidy up contract package
