+++ contract/population_types.go
@@ -4,13 +4,6 @@ type PopulationType struct {
 	Name string `json:"name"`
 }
 
-func NewPopulationType(name string) *PopulationType {
-	return &PopulationType{Name: name}
-}
-
-type GetPopulationTypesResponse struct {
-	PopulationTypes
-}
 type PopulationTypes struct {
 	Items []PopulationType `json:"items"`
 }
@@ -18,7 +11,11 @@ type PopulationTypes struct {
 func NewPopulationTypes(names []string) PopulationTypes {
 	items := make([]PopulationType, len(names))
 	for i, name := range names {
-		items[i] = *NewPopulationType(name)
+		items[i] = PopulationType{name}
 	}
 	return PopulationTypes{Items: items}
+}
+
+type GetPopulationTypesResponse struct {
+	PopulationTypes
 }
